Prevent printing RabbitMQ username/password to logs - not a major security issue as only admin can read the logs but better practice
